Fix 400 Bad Request: seed overflow Int, add action sync and AI turn endpoint
The seed field (initialized from Date.now()) exceeded Java Int.MAX_VALUE,
causing Jackson deserialization to fail on every /api/game/action request.
Changed seed to Long in SerializedGameState and GameState, with .toInt()
at noise/combat function boundaries.

Also fixes action sync sending post-action state instead of pre-action
state (which would double-apply the reducer), adds server-side AI turn
endpoint with OpenAPI spec, error logging advice, and test providers for
HttpClient in specs.

+++ server/src/main/kotlin/com/stellarhex/ai/AiService.kt
@@ -90,7 +90,7 @@ object AiService {
             // Try attack
             if (unit.weapon != null && !unit.hasAttacked) {
                 val attackResult = AiScoring.scoreAttack(
-                    unit, visibleEnemies, current.seed + current.turn, visibleEnemyBuildings,
+                    unit, visibleEnemies, (current.seed + current.turn).toInt(), visibleEnemyBuildings,
                 )
                 if (attackResult != null) {
                     val target = current.units[attackResult.targetId]
@@ -144,7 +144,7 @@ object AiService {
                             val cost = pathCost(path, hexLookup, costOverride)
                             current = GameReducer.reduce(
                                 current,
-                                GameAction.MoveUnit(freshUnit.id, path, cost),
+                                GameAction.MoveUnit(freshUnit.id, path, cost.toDouble()),
                             )
 
                             // After moving, try to collect anomaly at destination
